requirements.txt
python-nmap
openpyxl